Move ext-intl from a requirement to a suggestion
ext-intl isn't a common extension to have installed, and for novices getting this extension enabled can be difficult. Since it's only required for the offline geocoder, I have moved it from a hard requirement which would break the install, to a suggestion which will warn the user if they don't have it.