Hungary Code Changes
In coming metadata release: National prefix (06) will be present by default in national format for both fixed line and mobile numbers based on Wikipedia. So making the code ready.

For Hungary, there is special logic in formatNumberForMobileDialing() to manually add national prefix as earlier we are not adding based on metadata. So we are removing it as part of this change; and it is no more needed. Also removing the relevant test data and unit tests as because this is the only case tested for Hungary. More details in cl/265831383

Evidence: https://en.wikipedia.org/wiki/Telephone_numbers_in_Hungary

Mimics: https://github.com/google/libphonenumber/pull/2387